requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-core
langchain-openai
faiss-cpu
sentence-transformers